[CLIENT-4200] Performance fix (#185)
* Replace `RwLock` to `ArcLock`

+++ tools/benchmark/src/main.rs
@@ -52,7 +52,7 @@ async fn main() {
 
 async fn connect(options: &Options) -> Client {
     let mut policy = ClientPolicy::default();
-    policy.conn_pools_per_node = options.conn_pools_per_node;
+    policy.conn_pools_per_node = options.conn_pools_per_node as u8;
     Client::new(&policy, &options.hosts).await.unwrap()
 }
